Don't expect everyone to support HEAD.
Too often, they don't. The most recent problematic example I've
encountered is https://support.microsoft.com/en-gb/kb/841290, which
redirects forever when you HEAD it.

This probably won't introduce failures for anyone; we're not reading the
content of the request so it won't download massive files. The only
times I've overridden this function is to make it use GET with known-bad
servers.