Don't inherit from TestCase at runtime.
Fixes an issue where importing any of the mixins would cause a bunch of
failing tests to run in any Django project that has views. This is why I
yanked v1.2.0 from PyPI.

Without python/typing#213, I don't think it's possible for us to get
mypy to enforce that InstantCoverageAPI is used as a TestCase mixin. I
couldn't convince mypy to be okay with calls to super().setUp(), for
instance.